requirements.txt
# Absolute Essential Dependencies for First Iteration

# Data manipulation
pandas>=2.0.0
openpyxl>=3.1.0

# Database connection
psycopg2-binary>=2.9.5

# Configuration management
python-dotenv>=1.0.0
PyYAML>=6.0.0
